do on-the-fly tile changing correctly.
if enabled, draw lower resolution behind higher resolution so that you
can see low res before the higher res is downloaded. this makes things a
little slow in webgl (60 fps when zoomed out, down to 30-40 when zoomed
in). but it works fine in sw (50-60 all around).

if we disable drawing lower resolution, sw is very fast (100+fps), and
webgl is around 50-60. but then you get interrupted with black areas
while zooming in, and maybe affine transformation artifacts.